+++ R2U2_SW/R2U2_C/src/AT/filter_movavg.h
@@ -44,6 +44,6 @@ typedef struct
  * initially the average of the number of included elements is calculated
  * once the windows size has been reached, the average is calculated over the whole window
  * */
-void filter_movavg_update_data(movAvg_t *pMovAvg, int16_t data);
+void filter_movavg_update_data(movAvg_t *pMovAvg, int32_t data);
 float filter_movavg_get(movAvg_t *pMovAvg);
 #endif
